requirements_.txt
# Core dependencies for the Streamlit application
streamlit==1.32.0           # Web framework for creating the interface
pandas==2.2.0              # Data manipulation and CSV handling
qrcode==7.4.2             # QR code generation for PIX payments
pillow==10.2.0            # Image processing (required by qrcode)
python-dotenv==1.0.0      # Environment variable management for security

# Optional dependencies for enhanced features
# Uncomment these as you implement additional features:
# sendgrid==6.11.0        # For sending confirmation emails
# gspread==6.0.0          # For Google Sheets integration
# oauth2client==4.1.3     # For Google Sheets authentication
# streamlit-authenticator==0.3.1  # For admin dashboard authentication